+++ CHANGELOG.md
@@ -4,10 +4,11 @@ Change Log
 4.10.0
 ------
 
+- New `SmallVec` support in confined collections via `smallvec` feature
+- New `IndexSet` and `IndexMap` support in confined collections via `indexmap` feature
 - New `CursorDeque` implementation in `io` module, extending over smart pointers wrapping `VecDeque`
 - Updated CI: added support for ubuntu-24.04.arm runners
 - General chore and clippy lint fixes
-- confinement: add `IndexSet` and `IndexMap` support via `indexmap` feature
 
 4.9.0
 -----